Move some responsibilities from renderer to render storage
Took 18 minutes
