+++ crates/dream-renderer/src/lib.rs
@@ -28,5 +28,6 @@ pub mod model;
 pub mod path_not_found_error;
 pub mod pbr_bind_groups_and_layouts;
 pub mod render_map_key;
+pub mod render_storage;
 pub mod renderer;
 pub mod texture;
